fix(derive): refactor span batch tx types
