+++ crates/derive/src/types/batch/span_batch/tx_data/mod.rs
@@ -1,0 +1,13 @@
+//! Contains all the Span Batch Transaction Data types.
+
+mod wrapper;
+pub use wrapper::SpanBatchTransactionData;
+
+mod legacy;
+pub use legacy::SpanBatchLegacyTransactionData;
+
+mod eip1559;
+pub use eip1559::SpanBatchEip1559TransactionData;
+
+mod eip2930;
+pub use eip2930::SpanBatchEip2930TransactionData;
